requirements.txt
transformers
torch
opencv-python
gtts
rich
openai-whisper
sounddevice
scipy
pillow
requests